add upcast_softmax to prior_transformer

+++ src/diffusers/pipelines/shap_e/renderer.py
@@ -631,7 +631,10 @@ class ShapERenderer(ModelMixin, ConfigMixin):
         # 2. Get the points along the ray and query the model
         directions = torch.broadcast_to(direction.unsqueeze(-2), [batch_size, *ts_shape, 3])
         positions = origin.unsqueeze(-2) + ts * directions
-
+        
+        directions = directions.to(self.mlp.dtype)
+        positions = positions.to(self.mlp.dtype)
+        
         optional_directions = directions if render_with_direction else None
 
         model_out = self.mlp(
